Fix COPY FROM SEGMENT segment rows completed count
The primary segment backends return an int64 count back to the
dispatcher but the dispatcher temporarily stores that int64 count in
an int32 variable.

You can clearly see this in a debugger:
-> 609                          if (res->numCompleted > 0)
                                    ^
   610                                  segment_rows_completed = res->numCompleted;
(lldb) p res->numCompleted
(int64) $3 = 2000
(lldb) p segment_rows_completed
(int) $4 = 0

If res->numCompleted is bigger than INT32_MAX, segment_rows_completed
will become negative and a check later in this function will reject
this negative number from being added to the total rows completed
count.

+++ src/backend/cdb/cdbcopy.c
@@ -440,7 +440,7 @@ processCopyEndResults(CdbCopy *c,
 	PGresult   *res;
 	struct pollfd	*pollRead = (struct pollfd *) palloc(sizeof(struct pollfd));
 	int			segment_rows_rejected = 0;	/* num of rows rejected by this QE */
-	int			segment_rows_completed = 0; /* num of rows completed by this
+	int64			segment_rows_completed = 0; /* num of rows completed by this
 											 * QE */
 
 	for (seg = 0; seg < size; seg ++)
